Complete favorites API task

+++ src/main/java/ru/moonlightapp/backend/storage/repository/content/ProductRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import ru.moonlightapp.backend.model.attribute.ProductType;
 import ru.moonlightapp.backend.storage.model.content.Product;
 
-public interface ProductRepository extends JpaRepository<Product, Long>, JpaSpecificationExecutor<Product> {
+public interface ProductRepository extends JpaRepository<Product, Integer>, JpaSpecificationExecutor<Product> {
 
     @Query("select min(p.price) from Product p where p.type = ?1")
     float findMinPrice(ProductType productType);
